modal and refactor
